+++ modules/tinymce/src/themes/silver/test/ts/browser/editor/SilverEditorTest.ts
@@ -145,6 +145,9 @@ describe('browser.tinymce.themes.silver.editor.SilverEditorTest', () => {
                 classes: [ arr.has('tox-editor-header') ],
                 children: [
                   s.element('div', {
+                    classes: [ arr.has('tox-spap') ]
+                  }),
+                  s.element('div', {
                     classes: [ arr.has('tox-menubar') ],
                     attrs: { role: str.is('menubar') },
                     children: [