+++ modules/tinymce/src/themes/silver/test/ts/browser/editor/SilverFixedToolbarContainerTest.ts
@@ -46,6 +46,9 @@ describe('browser.tinymce.themes.silver.editor.SilverFixedToolbarContainerTest',
                     classes: [ arr.has('tox-editor-header') ],
                     children: [
                       s.element('div', {
+                        classes: [ arr.has('tox-spap') ]
+                      }),
+                      s.element('div', {
                         classes: [ arr.has('tox-menubar') ],
                         attrs: { role: str.is('menubar') },
                         children: [