+++ modules/tinymce/src/themes/silver/test/ts/module/StickyHeaderUtils.ts
@@ -17,6 +17,9 @@ const staticPartsInner = (s: ApproxStructure.StructApi, _str: ApproxStructure.St
   // should not change
   [
     s.element('div', {
+      classes: [ arr.has('tox-spap') ]
+    }),
+    s.element('div', {
       classes: [ arr.has('tox-menubar') ]
     }),
     s.element('div', {}),